Completely automate maven dependencies extraction
Replace the current script which just extracts the dependencies from a
build log with a new deno script which handles the whole thing,
including download, extraction, untaring, etc.

We use deno here because it has a solid story for 3rd party dependencies
in single-file scripts, with its URL and JSR imports, and because I like
typescript a lot more than Python.